+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Short HEVC + AC-3 test content
 - Generation of CMAF ingest streams with a REST-based API
 - New program `cmaf-ingest-receiver` that can receive one or more CMAF ingest streams
+- New option `--whitelistblocks` for unlimited number of requests to some CIDR blocks
 
 ### Fixed